UTF-8 is not Unicode
UTF-8 is an encoding from Unicode characters to bytes; a UTF-8 string
is made out of bytes and will not itself be accepted by the library.
If you have a UTF-8 string, you must first decode it from UTF-8 into
Unicode before passing it to the library.

Discussed in the comments at
https://github.com/EliFinkelshteyn/alphabet-detector/commit/67120416591879293c7a23a3d3430b4b41f755e4

Signed-off-by: Anders Kaseorg <andersk@mit.edu>